requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
scikit-learn
xgboost
plotly
streamlit
requests
python-dotenv
openai
